validate: move from plugin.api.validate package
- Move `streamlink.plugin.api.validate` to `streamlink.validate`:
  The validate sub-package doesn't belong into into the `plugin.api`
  sub-package, as it's not only being used by Streamlink's plugins,
  but also by the webbrowser API. It also has useful applications in
  other components of Streamlink, which has previously been avoided
  because of the sub-package layout.
- Keep re-exports in the `streamlink.plugin.api.validate` wrapper module
  where plugins should always import from, similar to the useragents
  wrapper module which re-exports from `session.http_useragents`.
- Update imports in affected (test-)modules
- Update references in API and API guide docs
- Add public interface note to API and API guide docs
