+++ src/streamlink/plugin/api/validate.py
@@ -1,0 +1,2 @@
+# noinspection PyUnresolvedReferences
+from streamlink.validate import *  # noqa: F403
