+++ src/streamlink/validate/_exception.py
