+++ src/streamlink/validate/_schemas.py
